moved commong filtering into FilterUtility Fixed issue in StreamFilter
