+++ graphql-api/src/main/java/com/expediagroup/streamplatform/streamregistry/graphql/filters/ProducerFilter.java
@@ -1,5 +1,5 @@
 /**
- * Copyright (C) 2018-2019 Expedia, Inc.
+ * Copyright (C) 2018-2020 Expedia, Inc.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -16,7 +16,7 @@
 package com.expediagroup.streamplatform.streamregistry.graphql.filters;
 
 import static com.expediagroup.streamplatform.streamregistry.graphql.filters.FilterUtility.matches;
-import static com.expediagroup.streamplatform.streamregistry.graphql.filters.SpecificationMatchUtility.matchesSpecification;
+import static com.expediagroup.streamplatform.streamregistry.graphql.filters.FilterUtility.matchesSpecification;
 
 import java.util.function.Predicate;
