+++ graphql-api/src/main/java/com/expediagroup/streamplatform/streamregistry/graphql/filters/StreamFilter.java
@@ -1,5 +1,5 @@
 /**
- * Copyright (C) 2018-2019 Expedia, Inc.
+ * Copyright (C) 2018-2020 Expedia, Inc.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -16,7 +16,8 @@
 package com.expediagroup.streamplatform.streamregistry.graphql.filters;
 
 import static com.expediagroup.streamplatform.streamregistry.graphql.filters.FilterUtility.matches;
-import static com.expediagroup.streamplatform.streamregistry.graphql.filters.SpecificationMatchUtility.matchesSpecification;
+import static com.expediagroup.streamplatform.streamregistry.graphql.filters.FilterUtility.matchesSchemaKey;
+import static com.expediagroup.streamplatform.streamregistry.graphql.filters.FilterUtility.matchesSpecification;
 
 import java.util.function.Predicate;
 
@@ -52,6 +53,7 @@ public class StreamFilter implements Predicate<Stream> {
       }
     }
 
-    return matchesSpecification(stream.getSpecification(), specQuery);
+    return matchesSchemaKey(stream.getSchemaKey(), schemaQuery)
+        && matchesSpecification(stream.getSpecification(), specQuery);
   }
 }
